Update security-requirements.md
Remove: "null" or
We encountered a persistent 401 error while implementing the code (given below) for a customer using Okta for authentication. Despite verifying the compatibility of the Outlook version with actionable messages and raising a support ticket with Microsoft, their suggestion to set it to an empty string. "null" seems confusing. To provide more clarity and convenience for developers, it would be helpful to include a valid example for "null" that can be easily copied and pasted into the code, or simply remove it. Thank you for your assistance!
/* -------- Code ----------- */
"actions": [
    {
        "type": "Action.Http",
        "id": "83667e35-2ef1-ab00-eb3a-9d9fa7ff0977",
        "title": "Submit",
        "url": "#ActionHttp#",
        "method": "POST",
        "body": "{\n  \"response\": \"Reject\",\n  \"comment\": \"{{commentReject.value}}\"\n}",
        "headers": [
            {
                "name": "Authorization"
            },
            {
                "name": "Content-type",
                "value": "application/json"
            }
        ]
    }
]